refactor: move resolution filter to `InstallOptions`
