+++ crates/uv-configuration/Cargo.toml
@@ -13,12 +13,14 @@ license = { workspace = true }
 workspace = true
 
 [dependencies]
+distribution-types = { workspace = true }
 pep508_rs = { workspace = true, features = ["schemars"] }
 platform-tags = { workspace = true }
 pypi-types = { workspace = true }
 uv-auth = { workspace = true }
 uv-cache = { workspace = true }
 uv-normalize = { workspace = true }
+uv-workspace = { workspace = true }
 
 clap = { workspace = true, features = ["derive"], optional = true }
 either = { workspace = true }
