+++ crates/uv/src/commands/project/sync.rs
@@ -1,9 +1,6 @@
 use anyhow::{Context, Result};
-use distribution_types::Name;
 use itertools::Itertools;
 use pep508_rs::MarkerTree;
-use rustc_hash::FxHashSet;
-use tracing::debug;
 use uv_auth::store_credentials_from_url;
 use uv_cache::Cache;
 use uv_client::{Connectivity, FlatIndexClient, RegistryClientBuilder};
@@ -194,8 +191,8 @@ pub(super) async fn do_sync(
     // Read the lockfile.
     let resolution = lock.to_resolution(project, markers, tags, extras, &dev)?;
 
-    // Set install-specific options.
-    let resolution = apply_install_options(&install_options, resolution, project);
+    // Filter resolution based on install-specific options.
+    let resolution = install_options.filter_resolution(resolution, project);
 
     // Add all authenticated sources to the cache.
     for url in index_locations.urls() {
@@ -283,66 +280,4 @@ pub(super) async fn do_sync(
     .await?;
 
     Ok(())
-}
-
-fn apply_install_options(
-    install_options: &InstallOptions,
-    resolution: distribution_types::Resolution,
-    project: &VirtualProject,
-) -> distribution_types::Resolution {
-    // If `--no-install-project` is set, remove the project itself.
-    let resolution =
-        apply_no_install_project(install_options.no_install_project, resolution, project);
-
-    // If `--no-install-workspace` is set, remove the project and any workspace members.
-    let resolution =
-        apply_no_install_workspace(install_options.no_install_workspace, resolution, project);
-
-    // If `--no-install-package` is provided, remove the requested packages.
-    apply_no_install_package(&install_options.no_install_package, resolution)
-}
-
-fn apply_no_install_project(
-    no_install_project: bool,
-    resolution: distribution_types::Resolution,
-    project: &VirtualProject,
-) -> distribution_types::Resolution {
-    if !no_install_project {
-        return resolution;
-    }
-
-    let Some(project_name) = project.project_name() else {
-        debug!("Ignoring `--no-install-project` for virtual workspace");
-        return resolution;
-    };
-
-    resolution.filter(|dist| dist.name() != project_name)
-}
-
-fn apply_no_install_workspace(
-    no_install_workspace: bool,
-    resolution: distribution_types::Resolution,
-    project: &VirtualProject,
-) -> distribution_types::Resolution {
-    if !no_install_workspace {
-        return resolution;
-    }
-
-    let workspace_packages = project.workspace().packages();
-    resolution.filter(|dist| {
-        !workspace_packages.contains_key(dist.name()) && Some(dist.name()) != project.project_name()
-    })
-}
-
-fn apply_no_install_package(
-    no_install_package: &[PackageName],
-    resolution: distribution_types::Resolution,
-) -> distribution_types::Resolution {
-    if no_install_package.is_empty() {
-        return resolution;
-    }
-
-    let no_install_packages = no_install_package.iter().collect::<FxHashSet<_>>();
-
-    resolution.filter(|dist| !no_install_packages.contains(dist.name()))
 }
